Moved the code used to build form hierarchy to one package
