+++ collect_app/src/main/AndroidManifest.xml
@@ -140,8 +140,8 @@ the specific language governing permissions and limitations under the License.
         <activity
             android:name=".preferences.screens.ProjectPreferencesActivity"
             android:theme="@style/Theme.Collect.Settings" />
-        <activity android:name=".activities.FormHierarchyActivity" />
-        <activity android:name=".activities.ViewOnlyFormHierarchyActivity" />
+        <activity android:name=".formhierarchy.FormHierarchyActivity" />
+        <activity android:name=".formhierarchy.ViewOnlyFormHierarchyActivity" />
         <activity
             android:name="org.odk.collect.geo.geopoint.GeoPointActivity"
             android:theme="@style/Theme.Collect"
